requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
python-dateutil==2.8.2
    # via -r requirements.in
pytz==2021.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
